requirements.txt
# LangChain and LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.1.18
langgraph>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Pydantic for schemas
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: for better console output
rich>=13.0.0


